# docs/api/openapi-v0.1.yaml
openapi: 3.0.3
info:
  title: Mneme API
  version: 0.1.0
  description: Frozen v0.1 contract. Implemented routes are verified against FastAPI-generated OpenAPI.
servers:
  - url: http://localhost:8000/v1
security:
  - demoToken: []
paths:
...
  /papers:
    get:
      operationId: listPapers
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Cursor-paginated papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperPage"
        default:
          $ref: "#/components/responses/Error"
...
  /papers/{paper_id}/summary:
    get:
      operationId: getPaperSummary
      parameters:
        - $ref: "#/components/parameters/PaperId"
      responses:
        "200":
          description: Ready summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "202":
          description: Summary generation accepted or in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        default:
          $ref: "#/components/responses/Error"
...
  /events:
    post:
      operationId: ingestEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 500
              items:
                $ref: "#/components/schemas/UserEvent"
      responses:
        "200":
          description: Idempotent ingestion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventIngestionResult"
        default:
          $ref: "#/components/responses/Error"
...
  /qa/ask:
    post:
      operationId: askQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Question"
      responses:
        "200":
          description: Grounded answer with source-matched citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    demoToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-demo-token
  parameters:
    PaperId:
      name: paper_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor for stable resource-specific keyset pagination
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Paper:
      type: object
      required: [id, arxiv_id, title, authors, abstract, primary_category, categories, pdf_url, processing_status, published_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        arxiv_id: {type: string}
        title: {type: string}
        authors: {type: array, items: {type: string}}
        abstract: {type: string}
        primary_category: {type: string}
        categories: {type: array, items: {type: string}}
        pdf_url: {type: string, format: uri}
        source_license: {type: string, nullable: true}
        published_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        processing_status:
          type: string
          enum: [metadata_only, queued, processing, ready, partial, failed]
    PaperPage:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Paper"}}
        next_cursor: {type: string, nullable: true}
    Summary:
      type: object
      required: [paper_id, status, tldr, source_match_status]
      properties:
        paper_id: {type: string, format: uuid}
        status: {type: string, enum: [ready, partial]}
        tldr: {type: string}
        key_claims: {type: array, items: {type: string}}
        methodology: {type: string}
        limitations: {type: string}
        source_match_status: {type: string, enum: [matched, partial, not_checked, unmatched]}
        claims: {type: array, items: {$ref: "#/components/schemas/SourcedClaim"}}
    SourcedClaim:
      type: object
      required: [text, matched]
      properties:
        text: {type: string}
        matched: {type: boolean}
        source:
          $ref: "#/components/schemas/ClaimProvenance"
          nullable: true
    ClaimProvenance:
      type: object
      required: [chunk_id, chunk_index, excerpt]
      properties:
        chunk_id: {type: string, format: uuid}
        chunk_index: {type: integer, minimum: 0}
        section_title: {type: string, nullable: true}
        page_start: {type: integer, nullable: true}
        page_end: {type: integer, nullable: true}
        excerpt: {type: string, maxLength: 300}
...
    UserEvent:
      type: object
      required: [event_id, event_type, occurred_at]
      properties:
        event_id: {type: string, format: uuid}
        event_type:
          type: string
          enum: [paper_impression, paper_opened, paper_saved, paper_skipped, paper_shared, question_asked, digest_dismissed]
        paper_id: {type: string, format: uuid, nullable: true}
        occurred_at: {type: string, format: date-time}
        duration_ms: {type: integer, minimum: 0, maximum: 2147483647, nullable: true}
        context: {type: object, additionalProperties: true}
    EventIngestionResult:
      type: object
      required: [accepted, duplicates]
      properties:
        accepted: {type: integer}
        duplicates: {type: integer}
...
    Question:
      type: object
      required: [question, paper_id]
      properties:
        question: {type: string, minLength: 1, maxLength: 500}
        paper_id: {type: string, format: uuid}
        conversation_id: {type: string, format: uuid, nullable: true}
    Citation:
      type: object
      required: [paper_id, section_title, source_match]
      properties:
        paper_id: {type: string, format: uuid}
        arxiv_id: {type: string}
        section_title: {type: string}
        chunk_id: {type: string, format: uuid}
        source_match: {type: boolean}
    Answer:
      type: object
      required: [answer, citations, source_match_status, conversation_id]
      properties:
        answer: {type: string}
        citations: {type: array, items: {$ref: "#/components/schemas/Citation"}}
        source_match_status: {type: string, enum: [matched, partial, insufficient_evidence]}
        conversation_id: {type: string, format: uuid}
...
    Job:
      type: object
      required: [id, stage, status]
      properties:
        id: {type: string, format: uuid}
        stage: {type: string}
        status: {type: string, enum: [queued, running, succeeded, failed]}
        error_code: {type: string, nullable: true}
        updated_at: {type: string, format: date-time}
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string}
        details: {type: object, additionalProperties: true}